fix shellcheck rm trendanalysis/tmp/

+++ bin/trendAnalyse.sh
@@ -87,7 +87,7 @@ function processRawdataToDB() {
 
 	log4Bash 'INFO' "${LINENO}" "${FUNCNAME:-main}" '0' "removing files from ${CHRONQC_TMP}"
 
-	rm -rf "${CHRONQC_TMP}"/*
+	rm -rf "${CHRONQC_TMP:?}"/*
 
 	if [[ -e "${PROCESSRAWDATATODB_CONTROLE_FILE_BASE}.finished" ]]
 	then
